Renamed the root namespace from ConsoleApplication4 to StudentCommandLineWebStore to match the repository name.
