+++ StudentCommandLineWebStore/Models/Addresses.cs
@@ -3,9 +3,9 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using ConsoleApplication4.Models.Types;
+using StudentCommandLineWebStore.Models.Types;
 
-namespace ConsoleApplication4.Models
+namespace StudentCommandLineWebStore.Models
 {
     class Addresses
     {
