+++ StudentCommandLineWebStore/Program.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace ConsoleApplication4
+namespace StudentCommandLineWebStore
 {
     class Program
     {
