+++ StudentCommandLineWebStore/Properties/AssemblyInfo.cs
