+++ StudentCommandLineWebStore/StudentCommandLineWebStore.csproj
